Solution: 771. Jewels and Stones
The brute-force solution would be for each character S[i], check each character J[j] in J to match it. If there's a match, count it as a jewel and halt the inner loop to go back to the next character S[i+1] and do the same. Finally, we have the number of jewels in counter. its time complexity is O(n^2).

On the other hand, if we use mapping, the solution can be found on O(n) time complexity. 
The idea is to map for each character J[i] to 1 such as map[J[i]]=1, then for each character S[i] just check the condition map[S[i]]==1, if the condition is true count it as a jewel otherwise ignore it. Therefore, we get the final result in counter.