Solution: 938. Range Sum of BST
It's a brute-force approach using DFS in which we traverse each node of the tree and check if it is inside [L, R]. Then add the node value to sum in case the checking condition holds true otherwise we proceed for the next node. Finally, we get our result in sum. The time complexity would be O(n) as we are traversing n number of nodes.