Solution: 1021. Remove Outermost Parentheses
The idea is to increase the count by 1 when we get S[i] = '(' and decrease the count by 1 in case S[i] = ')'. When the count = 0, we do not take this S[i] into out result string but the trick is when S[i] = '(', check the count before increasing it and in case of S[i] = ')', check the count after decreasing it. Therefore, we get the resulted string in str;