Solution: 804. Unique Morse Code Words
This is also a very easy problem to solve. First, we should create a map for each latter to Morse Code. Then for each word, we construct Morse Code Word and store it in the set data structure which removes the duplicate words and keeps only unique words. So the size of this set data structure is the final answer to the problem.