Solution: 905. Sort Array By Parity
In this problem, the built-in sort function of C++ is used with a custom comparator. The comparator returns true when an even number sits before an odd number such as [4,3] and false otherwise. When the sort function gets true from comparator function it does not swap the comparative two values but it operates a swap if the comparator returns false. In our case, for instance [4,3], a = 4%2 = 0 and b = 3%2 = 1. So the comparator returns true for such example and the sort function does nothing to them. Since the sort function of C++ takes O(nlogn) time, so the time complexity of this solution is also O(nlogn).