Solution: 657. Robot Return to Origin
To return to the origin, the Robot needs to have U = D and R = L. Otherwise it is not possible to come back to the origin after starting from it. So first, we check if the total number of moves is odd then we should not check further. The time complexity is O(n) where n is the total number of moves.